VB-4000 - Fix issue where unapproved visitors were not displaying correctly
Bug: On the prisoner profile, if a visitor had been unapproved, the past
visits they attended would have their name as "undefined". This was due
to our API only returning approved visitors.

Fix / Rework: Change the orchestration to use a new endpoint on the
prisoner-contact-registry which takes a flag "approvedVisitorsOnly".

The prisoner profile will pass in false for this endpoint and other
calls will pass in true to preserve original functionality.