VB-4121 - Replace prison API with Alerts API (#276)
* VB-4121 - Replace prison API with Alerts API

Encountered a NPE where description field was null, adding nullable ?
to field.

* add json property to map to API field being returned

* Fix deserialisation issue with fields prefixed with "is"

on the Alerts API call, the field returned is named isActive,
on our side it was set as "active". I added a JSONProperty to fix
this mapping however, this caused tests to fail with deserialising
problems. To fix this, i added a kotlinModule() to the object mapper.